amd64: Buildable port; builds without errors but can't be run

+++ kernel/includes/cedille/pmm.h
@@ -10,21 +10,21 @@
 ///Init the pmm
 void init_pmm();
 ///Set the frame that governs address to used
-void pmm_set_frame(uint32_t address);
+void pmm_set_frame(uintptr_t address);
 ///Set the frame that governs address to free
-void pmm_clear_frame(uint32_t address);
+void pmm_clear_frame(uintptr_t address);
 ///Is the frame that governs address used?
-uint32_t pmm_test_frame(uint32_t address);
+uintptr_t pmm_test_frame(uintptr_t address);
 ///First free frame
-uint32_t pmm_first_frame();
+uintptr_t pmm_first_frame();
 ///How many possible frames?
-uint32_t pmm_frame_amount();
+uintptr_t pmm_frame_amount();
 ///Allocate a frame from page
-void pmm_alloc_frame(uint32_t address, int kernel, int rw);
-void pmm_shim_alloc_frame(uint32_t address, int kernel, int rw);
+void pmm_alloc_frame(uintptr_t address, int kernel, int rw);
+void pmm_shim_alloc_frame(uintptr_t address, int kernel, int rw);
 ///Free a frame
-void pmm_free_frame(uint32_t address);
-void pmm_shim_free_frame(uint32_t address);
+void pmm_free_frame(uintptr_t address);
+void pmm_shim_free_frame(uintptr_t address);
 
 void pmm_set_maxmem(uintptr_t max);
 #endif
